fix(clerk-js): Proper check for ability to remove org image
