+++ packages/clerk-js/src/ui/components/OrganizationProfile/ProfileSettingsPage.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 
+import { isDefaultImage } from '../../../utils';
 import { useWizard, Wizard } from '../../common';
-import { useCoreOrganization } from '../../contexts';
+import { useCoreOrganization, useOptions } from '../../contexts';
 import { localizationKeys } from '../../customizables';
 import { ContentPage, Form, FormButtons, SuccessPage, useCardState, withCardStateProvider } from '../../elements';
 import { handleError, useFormControl } from '../../utils';
@@ -14,6 +15,7 @@ export const ProfileSettingsPage = withCardStateProvider(() => {
   const subtitle = localizationKeys('organizationProfile.profilePage.subtitle');
   const card = useCardState();
   const [avatarChanged, setAvatarChanged] = React.useState(false);
+  const { experimental_enableClerkImages } = useOptions();
   const { organization } = useCoreOrganization();
 
   const wizard = useWizard({ onNextStep: () => card.setError(undefined) });
@@ -86,7 +88,15 @@ export const ProfileSettingsPage = withCardStateProvider(() => {
           <OrganizationProfileAvatarUploader
             organization={organization}
             onAvatarChange={uploadAvatar}
-            onAvatarRemove={organization.logoUrl ? onAvatarRemove : null}
+            onAvatarRemove={
+              experimental_enableClerkImages
+                ? isDefaultImage(organization.experimental_imageUrl)
+                  ? onAvatarRemove
+                  : null
+                : organization.logoUrl
+                ? onAvatarRemove
+                : null
+            }
           />
           <Form.ControlRow elementId={nameField.id}>
             <Form.Control
