+++ packages/clerk-js/src/utils/index.ts
@@ -28,3 +28,4 @@ export * from './componentGuards';
 export * from './queryStateParams';
 export * from './decodeBase16';
 export * from './authPropHelpers';
+export * from './image';
