+++ packages/clerk-js/src/utils/user.ts
@@ -10,8 +10,7 @@ export const getFullName = ({ firstName, lastName, name }: NameHelperParams) =>
 export const getInitials = ({ firstName, lastName, name }: NameHelperParams) =>
   [(firstName || '')[0], (lastName || '')[0]].join('').trim() || (name || '')[0];
 
-import { isomorphicAtob } from '@clerk/shared';
-import type { EncodedImageData, UserResource } from '@clerk/types';
+import type { UserResource } from '@clerk/types';
 
 export const getIdentifier = (user: Partial<UserResource>): string => {
   if (user.username) {
@@ -31,20 +30,4 @@ export const getIdentifier = (user: Partial<UserResource>): string => {
   }
 
   return '';
-};
-
-export const isDefaultProfileImage = (url: string) => {
-  // Remove this check when renaming experimental_imageUrl to imageUrl
-  if ((url || '').includes('gravatar')) {
-    return true;
-  }
-
-  try {
-    const encoded = new URL(url).pathname.replace('/', '');
-    const decoded = isomorphicAtob(encoded);
-    const obj = JSON.parse(decoded) as EncodedImageData;
-    return obj.type === 'default';
-  } catch {
-    return false;
-  }
 };
